refactor(backend): delete lots of old, dead things
rearrange packages, removing the `update` package in the process. I'm not a fan
of how this thickens cmd/serve.go, but I think it's worth it to drop a level of
indrection by removing the update package & constructing an api service directly
