--- qrimatic/scheduler/testdata/test.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-echo "hello world!"
