--- qrimatic/update/testdata/hello.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-echo "hello"
